Fix: Switch -> PostingSwtich로 변경, usePosingStateMutation react-query hook 적용
